switch back to public subnets
when using private subnets with public EIPs, all outbound traffic goes
through the NAT gateway.  this seems to cause stateful problems with
inbound traffic using the EIPs, as traffic comes in the IGW but out the
NAT gateway.  clients external to the VPC see this as a TCP ACK reset.
switching back to public subnet and using security groups to block all
undesired incoming traffic works around this issue.